Backend/backend_requirements.txt
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
